+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Trace
 
 - *Breaking*: Big refactor to make the type decoding and value reading be structured instead of it all being strings
+- *Breaking*: Tracing has been made crossplatform with an implementation for Cortex-M
 - Added reading capability for tagged unions (fancy Rust enums)
 - Added colorized output
 - Made it so that transparent types can be added to lessen the clutter in the trace
@@ -26,6 +27,7 @@
 - *Breaking*: Simplified the MemoryRegion trait and made it fallible
 - *Breaking*: Simplified the RegisterData trait
 - *Breaking*: DeviceMemory now takes a `'memory` lifetime so that not all data has to be owned
+- *Breaking*: The FromIterator impl taking `&u8` has been removed leaving only one impl that takes `u8`. Just pass an iterator with `.copied()` to it.
 
 ## Capture 0.2.0 (03-05-22)